requirements.txt
pyspellchecker
deep-translator
Pillow
pygame